% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_countries.R
\name{fetch_countries}
\alias{fetch_countries}
\title{Fetch country listing}
\usage{
fetch_countries()
}
\description{
\code{fetch_countries} returns a dataframe of all countries with available DHS survey data and their associated DHS and ISO codes.
}
\seealso{
\code{\link{fetch_data}}
}
