% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_indicators.R
\name{fetch_indicators}
\alias{fetch_indicators}
\title{Fetch indicator listing}
\usage{
fetch_indicators()
}
\description{
\code{fetch_indicators} returns a dataframe of all available DHS survey indicators, e.g. indicator id, definition, and denominator.
}
\seealso{
\code{\link{fetch_data}}
}
