% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_tags.R
\name{fetch_tags}
\alias{fetch_tags}
\title{Fetch tag listing}
\usage{
fetch_tags()
}
\description{
\code{fetch_tags} returns a dataframe detailing DHS survey tags. Tags are useful when seeking to extract data by thematic area, e.g. immunization, pneumonia, or HIV.
}
\seealso{
\code{\link{fetch_data}}
}
