% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_api_key.R
\name{set_api_key}
\alias{set_api_key}
\title{Set DHS API key}
\usage{
set_api_key(dhs_api_key)
}
\arguments{
\item{dhs_api_key}{Input DHS API key as string. Please see \url{https://api.dhsprogram.com/#/introdevelop.html} for authentication details.}
}
\value{
Returns the currently set \code{dhs_api_key}.
}
\description{
\code{set_api_key} captures the DHS API key for use in \code{fetch_data}.
}
\examples{
set_api_key("ABCDEF-123456")
}
\seealso{
\code{\link{fetch_data}}
}
