% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_return_fields.R
\name{set_return_fields}
\alias{set_return_fields}
\title{Set return fields for DHS API}
\usage{
set_return_fields(dhs_return_fields)
}
\arguments{
\item{dhs_return_fields}{Input return fields to populate the retrieved dataframe from the DHS API. Please see \url{https://api.dhsprogram.com/rest/dhs/data/fields} for complete list of available fields.}
}
\value{
Captures a vector of the specified return fields.
}
\description{
\code{set_return_fields} captures a set of return fields for use in \code{fetch_data}.
}
\examples{
set_return_fields(c("Indicator", "IndicatorId","CountryName", "DHS_CountryCode",
                    "SurveyYear", "SurveyYearLabel", "SurveyType",
                    "CharacteristicCategory", "CharacteristicLabel", "ByVariableLabel",
                    "DenominatorWeighted", "DenominatorUnweighted",
                    "CILow", "CIHigh", "Value"))
}
\seealso{
\code{\link{fetch_data}}
}
